Rename Weather monitoring syatem output.jpg to Weather monitoring system output.jpg
This project was about designing and implementing a real-time temperature monitoring system using an analog sensor and microcontroller-based setup. The main goal was to measure temperature accurately, convert the sensor’s analog signal into digital form, and display the readings in real time on an LCD screen.

The system used the LM35 temperature sensor to sense temperature and produce a proportional analog voltage. Since the microcontroller (LPC2129) cannot directly read analog signals with high precision, an MCP3204 12-bit ADC was interfaced through the SPI protocol to perform analog-to-digital conversion. The LPC2129 received this digital temperature data, processed it to obtain the actual temperature value, and displayed it on a 16x2 LCD.

This project demonstrated how register-level programming can be used to configure SPI and GPIO modules in embedded systems. It also provided hands-on experience in hardware interfacing, real-time data acquisition, and embedded debugging using the Keil µVision IDE.

In essence, the project focused on real-time temperature measurement and display using SPI-based ADC communication and embedded C programming on the LPC2129 microcontroller.